billing_compose/django/requirements.txt
celery==5.2.7
django-cors-headers==3.12.0
django-filter==22.1
django-payments[stripe]==1.0.0
django-split-settings==1.2.0
Django==4.1.2
djangorestframework==3.12.4
drf-nested-routers==0.93.4
drf-spectacular==0.24.2
gunicorn==20.1.0
psycopg2-binary==2.9.4
PyJWT==2.4.0
python-dotenv==0.21.0
pytz==2022.4
redis==4.3.4
stripe==4.2.0
yookassa==2.3.2
python-dateutil==2.8.2
django-extensions==3.2.1
kafka-python==2.0.2
loguru==0.6.0